Fix listener and solicitation destination address
RFC4286, section 4.2.2 states that the IP destination address for
MRD solicitation messages has to be the All-Routers multicast address
and not the All-Snoopers one. Fixing this both on the receiver (mrdisc)
and transmitter (solicit) side:

The latter needs to transmit to the All-Routers multicast
address and the former needs to join and listen to it.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>