Fix IGMP checksum
RFC4286, section 3.2.3 says:

"For IPv4 it is the 16-bit one's complement of the one's
 complement sum of the IGMP message, starting with the Type field.
 For computing the checksum, the checksum field is set to 0."

Which means (and is a bit confusing, because these two sentences are in
reverse order time-wise):

Before computing the checksum, the IGMP checksum field needs to be set
to 0. Then the checksum is computed over the whole IGMP message,
including the IGMP checksum field. And then the IGMP checksum field is
set again to this newly calculated value.

For the receiver side, RFC4286, section 3.5 says:

"Upon receiving an Advertisement message, devices validate the message
 with the following criteria:

   1.  The checksum is correct
   [...]

An Advertisement not meeting the validity requirements MUST be
silently discarded [...]"

Which means a correct checksum is mandatory for a multicast router
advertisement message to be processed. A zero checksum is usually not
valid (in contrast to UDP).

Tested against the upcoming MRD implementation in the Linux bridge.

Fixes: 78096910242 ("Remove badly calculated UDP/IGMP checksum, not needed for IPv4")
Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>