Add IPv6 support to mrdisc and solicit
This adds IPv6 support. By default both IPv4 and IPv6 operations are
performed simultaneously. The user can revert to v4-only or v6-only by
using the newly introduced "-4" and "-6" switches.

One noteable difference observed on a Linux 4.19 and 5.1-rc3 kernel was
that contrary to IPv4/IGMP the IPv6/ICMPv6 stack in the kernel
calculates the ICMPv6 checksum on its own and does not need to be
calculated in userspace. (other operating systems were untested)

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>